Added BlindRod model Update all other models Add BlindRod controller Update Blind profile controller Update migrations Add Bottom Bar and Rod CRUD views Update Blind Profile CRUD views Fix Blind Form and Blind View components to update model Move workings component to index

+++ Shared/BlindRailing.cs
@@ -4,6 +4,6 @@
     {
         public int Id { get; init; }
         public string Name { get; set; } = "Default Railing";
-        public float Height { get; set; } = 2.5f;
+        public double Height { get; set; } = 2.5f;
     }
 }
